**Why**:
* When running `rpsec spec/ --profile`, i18n-tasks was slowest file
* Ran 3 times, each time file took ~2 seconds to run
* Upon investigation, this appears to relate to Simplecov
* See https://github.com/glebm/i18n-tasks/issues/221
* Also https://github.com/colszowka/simplecov/issues/404
* Since there is no proposed solution at the moment (slowness appears to
  be coming from `Coverage`, which is part of the Ruby stdlib), the
  short term solution is to only run Simplecov on Travis, where required
* New timing for i18n-tasks: under .2 seconds!!!

Before:

```
Top 10 slowest example groups:
  I18n
    1.8 seconds average (3.59 seconds / 2 examples) ./spec/i18n_spec.rb:4
  Users::ServiceProviders
    0.17479 seconds average (1.22 seconds / 7 examples) ./spec/requests/service_providers_spec.rb:3
```

After:

```
Top 10 slowest example groups:
  Users::ServiceProviders
    0.16671 seconds average (1.17 seconds / 7 examples) ./spec/requests/service_providers_spec.rb:3
  I18n
    0.14545 seconds average (0.29089 seconds / 2 examples) ./spec/i18n_spec.rb:4
```